# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Marchi API
  description: |
    JSON REST API for Marchi, a self-hosted email archiving service.

    All `/api/v1/*` endpoints (and `GET /ws`) require an authenticated browser
    session: `POST /unlock` with the Master Key password first, which sets a
    session cookie. Requests without a valid session receive `401`.

    State-changing requests (`POST`/`PUT`/`DELETE`) additionally require a
    CSRF token, issued as a `csrf_` cookie on any prior `GET` and echoed back
    via the `X-Csrf-Token` header.

    Error responses across the whole API are **plain text**, not JSON: the
    body is the raw error message, `Content-Type: text/plain; charset=utf-8`.
    This document models that as the `PlainTextError` response used on every
    non-2xx response.
  version: "0.1.0"
  license:
    name: MIT
servers:
  - url: /api/v1
security:
  - sessionCookie: []
...
paths:
...
  /emails/{id}/restore-logs:
    get:
      tags: [Restore]
      summary: Restore history for one email, newest first
      parameters: [{ $ref: "#/components/parameters/EmailID" }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/RestoreLog" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "500": { $ref: "#/components/responses/ServerError" }
  /rules:
    get:
      tags: [Rules]
      summary: List all rules
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Rule" }
        "500": { $ref: "#/components/responses/ServerError" }
    post:
      tags: [Rules]
      summary: Create a rule
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RuleRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Rule" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "500": { $ref: "#/components/responses/ServerError" }
  /rules/dry-run:
    post:
      tags: [Rules]
      summary: Test an unsaved condition tree against already-archived emails
      description: >-
        Read-only — never creates or modifies a rule. Reconstructs each
        scanned email's match candidate from persisted emails/attachments
        rows (no raw .eml re-read), so a rule (especially
        archive_and_delete) can be sanity-checked before it's ever saved.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DryRunRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DryRunResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "500": { $ref: "#/components/responses/ServerError" }
  /rules/{id}:
    put:
      tags: [Rules]
      summary: Replace a rule
      parameters: [{ $ref: "#/components/parameters/RuleID" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RuleRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Rule" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/ServerError" }
    delete:
      tags: [Rules]
      summary: Delete a rule
      parameters: [{ $ref: "#/components/parameters/RuleID" }]
      responses:
        "204": { description: Deleted }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/ServerError" }
...
  /restore:
    post:
      tags: [Restore]
      summary: Restore archived emails into a target account/folder
      description: >-
        Validated synchronously; the actual restore(s) run as a background
        job. Subscribe to `GET /ws` for progress (`type: "restore"` events
        carrying `succeeded`/`failed` counts).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RestoreRequest" }
      responses:
        "202":
          description: Accepted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobIDResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404":
          description: Target account not found
          content: { text/plain: { schema: { type: string } } }
        "500": { $ref: "#/components/responses/ServerError" }
  /restore/bulk:
    post:
      tags: [Restore]
      summary: Restore every email in an entire source account or a single source folder
      description: >-
        Unlike `POST /restore`, the caller doesn't select emails
        individually — every email in scope is resolved server-side. For
        `scope: "folder"`, every email lands directly in
        `target_folder_root`. For `scope: "account"`, emails are grouped
        by their source folder and each group gets its own subfolder
        under `target_folder_root` (so the source folder structure isn't
        flattened). Progress over `GET /ws` the same as `POST /restore`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BulkRestoreRequest" }
      responses:
        "202":
          description: Accepted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobIDResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404":
          description: Target account, source account, or source folder not found
          content: { text/plain: { schema: { type: string } } }
        "500": { $ref: "#/components/responses/ServerError" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session_id
      description: Set by `POST /unlock`. State-changing requests also need the `X-Csrf-Token` header (value from the `csrf_` cookie).
  parameters:
...
    EmailID:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
    RuleID:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
  responses:
    BadRequest:
      description: Invalid request (bad path param, malformed body, or a validation error)
      content: { text/plain: { schema: { type: string } } }
    NotFound:
      description: Not found
      content: { text/plain: { schema: { type: string } } }
    ServerError:
      description: Internal server error
      content: { text/plain: { schema: { type: string } } }
  schemas:
    JobIDResponse:
      type: object
      properties:
        job_id: { type: string }
      required: [job_id]
...
    RuleNode:
      type: object
      description: >-
        Either {op, children} (a group) or {type, value} (a leaf), never
        both. Groups nest up to 3 levels deep.
      properties:
        op: { type: string, enum: [and, or] }
        children:
          type: array
          items: { $ref: "#/components/schemas/RuleNode" }
        type:
          type: string
          enum:
            - from_contains
            - from_domain
            - from_exact
            - to_contains
            - to_domain
            - subject_contains
            - has_attachments
            - attachment_type
            - size_greater_than
            - size_less_than
            - date_after
            - date_before
            - folder_is
            - folder_is_not
            - account_is
        value: { type: string, description: "Always a string, regardless of the condition's semantic type." }
    Rule:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        priority: { type: integer }
        conditions: { $ref: "#/components/schemas/RuleNode" }
        action: { type: string, enum: [archive, skip, archive_and_delete, archive_and_mark_read] }
        is_active: { type: boolean }
    RuleRequest:
      type: object
      required: [name, conditions, action]
      properties:
        name: { type: string }
        priority: { type: integer }
        conditions: { $ref: "#/components/schemas/RuleNode" }
        action: { type: string, enum: [archive, skip, archive_and_delete, archive_and_mark_read] }
        is_active: { type: boolean }
    DryRunRequest:
      type: object
      required: [conditions]
      properties:
        conditions: { $ref: "#/components/schemas/RuleNode" }
        account_id: { type: integer, format: int64, nullable: true, description: "Scope the scan to one account; omit to scan all accounts." }
        limit: { type: integer, description: "Caps how many archived emails are scanned (default and max 5000)." }
    DryRunResponse:
      type: object
      properties:
        total_scanned: { type: integer }
        matched_count: { type: integer }
        samples:
          type: array
          description: "Up to 20 matched emails, for a representative preview."
          items:
            type: object
            properties:
              id: { type: integer, format: int64 }
              subject: { type: string }
              from: { type: string }
              date: { type: string }
...
    RestoreRequest:
      type: object
      required: [email_ids, target_account_id, target_folder]
      properties:
        email_ids: { type: array, items: { type: integer, format: int64 } }
        target_account_id: { type: integer, format: int64 }
        target_folder: { type: string }
    BulkRestoreRequest:
      type: object
      required: [scope, scope_id, target_account_id, target_folder_root]
      properties:
        scope: { type: string, enum: [account, folder] }
        scope_id: { type: integer, format: int64, description: "Source account id (scope=account) or source folder id (scope=folder)." }
        target_account_id: { type: integer, format: int64 }
        target_folder_root: { type: string }
    RestoreLog:
      type: object
      properties:
        id: { type: integer, format: int64 }
        email_id: { type: integer, format: int64 }
        target_account_id: { type: integer, format: int64 }
        target_folder: { type: string }
        method: { type: string, enum: [imap_append, smtp] }
        status: { type: string, enum: [completed, failed] }
        error_msg: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
...